requirements.txt
# Web framework
Flask>=3.0,<4.0
gunicorn>=21.2,<22.0

# Database
psycopg2-binary>=2.9,<3.0

# MCP client (SerpApi MCP server)
mcp>=1.0,<2.0

# Google Gemini AI
google-genai>=1.0

# HTTP client (Evolution API + SerpApi HTTP fallback)
httpx>=0.27,<1.0

# Environment
python-dotenv>=1.0,<2.0
